FXOS8700CQ: Added the option to select whether to apply low and high-pass filters to the Pulse Interrupt function.

+++ include/FXOS8700CQ.h
@@ -43,7 +43,7 @@ class FXOS8700CQ{
         void setNewData(FXO::Interrupt_Pin pin, void (*function)());
         void setAccelerationMagnitude(FXO::Interrupt_Pin pin, void (*function)(), uint8_t count, bool resetCount, uint8_t config, float threshold, float* reference);
         void setFreefallMotion(FXO::Interrupt_Pin pin, void (*function)(), float count, bool resetCount, uint8_t config, float threshold, float xThreshold=0, float yThreshold=0, float zThreshold=0);
-        void setPulse(FXO::Interrupt_Pin pin, void (*function)(), uint8_t config, float timing, float* threshold, float latency, float window);
+        void setPulse(FXO::Interrupt_Pin pin, void (*function)(), uint16_t config, float timing, float* threshold, float latency, float window);
         void setOrientation(FXO::Interrupt_Pin pin, void (*function)(), float count, bool resetCount, FXO::Lockout_Angle lockout, FXO::Orientation_Angle angle, FXO::Trip_Threshold threshold, FXO::Trip_Delta delta);
         void setTransient(FXO::Interrupt_Pin pin, void (*function)(), float count, bool resetCount, uint8_t config, float threshold);
         void setFIFO(FXO::Interrupt_Pin pin, void (*function)(), FXO::FIFO_Mode mode, uint8_t watermark = 0, uint8_t config = 0);
